Time: 3 ms (47.13%), Space: 6 MB (45.55%) - LeetHub

+++ 0231-power-of-two/0231-power-of-two.cpp
@@ -3,7 +3,7 @@ public:
     bool isPowerOfTwo(int n) {
         
         
-        int N = n;
+        long long int N = n;
         
         if(N<=0)
             return false;
